+++ src/Components/connection-center/src/Handler/AlwaysNew/AlwaysNewConnectionManagerStatistics.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Imi\ConnectionCenter\Handler\AlwaysNew;
 
+use Imi\ConnectionCenter\Contract\AbstractConnectionManagerStatistics;
+
 /**
  * 总是创建新连接管理器状态.
  */
-use Imi\ConnectionCenter\Contract\AbstractConnectionManagerStatistics;
-
 class AlwaysNewConnectionManagerStatistics extends AbstractConnectionManagerStatistics
 {
 }
