+++ src/Components/connection-center/src/Handler/AlwaysNew/AlwaysNewConnectionManagerWritableStatistics.php
@@ -4,38 +4,14 @@ declare(strict_types=1);
 
 namespace Imi\ConnectionCenter\Handler\AlwaysNew;
 
+use Imi\ConnectionCenter\Contract\TConnectionManagerWritableStatistics;
+
 /**
  * 总是创建新连接管理器可写状态.
  */
 class AlwaysNewConnectionManagerWritableStatistics extends AlwaysNewConnectionManagerStatistics
 {
-    public function addCreateConnectionTimes(): int
-    {
-        return ++$this->createConnectionTimes;
-    }
-
-    public function addGetConnectionTimes(): int
-    {
-        return ++$this->getConnectionTimes;
-    }
-
-    public function addReleaseConnectionTimes(): int
-    {
-        return ++$this->releaseConnectionTimes;
-    }
-
-    public function setGetConnectionTime(float $time): void
-    {
-        $this->lastGetConnectionTime = $time;
-        if ($time < $this->minGetConnectionTime)
-        {
-            $this->minGetConnectionTime = $time;
-        }
-        if ($time > $this->maxGetConnectionTime)
-        {
-            $this->maxGetConnectionTime = $time;
-        }
-    }
+    use TConnectionManagerWritableStatistics;
 
     public function toStatus(): AlwaysNewConnectionManagerStatistics
     {
